Rename duckdb/postgres extension
