+++ .github/workflows/ci-workflow.yml
@@ -497,13 +497,13 @@ jobs:
       - uses: actions/checkout@v3
 
       - name: Update PostgreSQL host
-        working-directory: extension/postgres_scanner/test/test_files
+        working-directory: extension/postgres/test/test_files
         env:
-          FNAME: postgres_scanner.test
+          FNAME: postgres.test
           FIND: "localhost"
         run: |
           node -e 'fs=require("fs");fs.readFile(process.env.FNAME,"utf8",(err,data)=>{if(err!=null)throw err;fs.writeFile(process.env.FNAME,data.replaceAll(process.env.FIND,process.env.PG_HOST),"utf8",e=>{if(e!=null)throw e;});});'
-          cat postgres_scanner.test
+          cat postgres.test
 
       - name: Install dependencies
         run: pip install rangehttpserver
@@ -533,13 +533,13 @@ jobs:
       - uses: actions/checkout@v3
 
       - name: Update PostgreSQL host
-        working-directory: extension/postgres_scanner/test/test_files
+        working-directory: extension/postgres/test/test_files
         env:
-          FNAME: postgres_scanner.test
+          FNAME: postgres.test
           FIND: "localhost"
         run: |
           node -e 'fs=require("fs");fs.readFile(process.env.FNAME,"utf8",(err,data)=>{if(err!=null)throw err;fs.writeFile(process.env.FNAME,data.replaceAll(process.env.FIND,process.env.PG_HOST),"utf8",e=>{if(e!=null)throw e;});});'
-          cat postgres_scanner.test
+          cat postgres.test
 
       - name: Install dependencies
         run: pip3 install --break-system-packages rangehttpserver
@@ -575,13 +575,13 @@ jobs:
         run: pip install rangehttpserver
 
       - name: Update PostgreSQL host
-        working-directory: extension/postgres_scanner/test/test_files
+        working-directory: extension/postgres/test/test_files
         env:
-          FNAME: postgres_scanner.test
+          FNAME: postgres.test
           FIND: "localhost"
         run: |
           node -e 'fs=require("fs");fs.readFile(process.env.FNAME,"utf8",(err,data)=>{if(err!=null)throw err;fs.writeFile(process.env.FNAME,data.replaceAll(process.env.FIND,process.env.PG_HOST),"utf8",e=>{if(e!=null)throw e;});});'
-          cat postgres_scanner.test
+          cat postgres.test
 
       - name: Extension test
         shell: cmd
