+++ Makefile
@@ -64,7 +64,7 @@ allconfig:
 	$(call config-cmake-release, \
 		-DBUILD_BENCHMARK=TRUE \
 		-DBUILD_EXAMPLES=TRUE \
-		-DBUILD_EXTENSIONS="httpfs;duckdb_scanner;postgres_scanner" \
+		-DBUILD_EXTENSIONS="httpfs;duckdb;postgres" \
 		-DBUILD_JAVA=TRUE \
 		-DBUILD_NODEJS=TRUE \
 		-DBUILD_PYTHON=TRUE \
@@ -79,7 +79,7 @@ alldebug:
 	$(call run-cmake-debug, \
 		-DBUILD_BENCHMARK=TRUE \
 		-DBUILD_EXAMPLES=TRUE \
-		-DBUILD_EXTENSIONS="httpfs;duckdb_scanner;postgres_scanner" \
+		-DBUILD_EXTENSIONS="httpfs;duckdb;postgres" \
 		-DBUILD_JAVA=TRUE \
 		-DBUILD_NODEJS=TRUE \
 		-DBUILD_PYTHON=TRUE \
@@ -162,7 +162,7 @@ example:
 
 extension-test:
 	$(call run-cmake-release, \
-		-DBUILD_EXTENSIONS="httpfs;duckdb_scanner;postgres_scanner" \
+		-DBUILD_EXTENSIONS="httpfs;duckdb;postgres" \
 		-DBUILD_EXTENSION_TESTS=TRUE \
 		-DENABLE_ADDRESS_SANITIZER=TRUE \
 	)
@@ -171,13 +171,13 @@ extension-test:
 
 extension-debug:
 	$(call run-cmake-debug, \
-		-DBUILD_EXTENSIONS="httpfs;duckdb_scanner;postgres_scanner" \
+		-DBUILD_EXTENSIONS="httpfs;duckdb;postgres" \
 		-DBUILD_KUZU=FALSE \
 	)
 
 extension-release:
 	$(call run-cmake-release, \
-		-DBUILD_EXTENSIONS="httpfs;duckdb_scanner;postgres_scanner" \
+		-DBUILD_EXTENSIONS="httpfs;duckdb;postgres" \
 		-DBUILD_KUZU=FALSE \
 	)
