+++ extension/CMakeLists.txt
@@ -2,19 +2,19 @@ if("httpfs" IN_LIST BUILD_EXTENSIONS)
     add_subdirectory(httpfs)
 endif()
 
-if ("duckdb_scanner" IN_LIST BUILD_EXTENSIONS)
+if ("duckdb" IN_LIST BUILD_EXTENSIONS)
     if(NOT __32BIT__)
         # DuckDB does not officially support 32-bit builds, so we disable the 
         # extension for 32-bit builds
-        add_subdirectory(duckdb_scanner)
+        add_subdirectory(duckdb)
     endif()
 endif()
 
-if ("postgres_scanner" IN_LIST BUILD_EXTENSIONS)
+if ("postgres" IN_LIST BUILD_EXTENSIONS)
     if(NOT __32BIT__)
         # DuckDB does not officially support 32-bit builds, so we disable the
         # extension for 32-bit builds
-        add_subdirectory(postgres_scanner)
+        add_subdirectory(postgres)
     endif()
 endif()
