+++ extension/duckdb/src/duckdb_table_catalog_entry.cpp
