+++ extension/duckdb/src/include/duckdb_table_catalog_entry.h
