+++ extension/duckdb/test/duckdb_database/other.db
