--- extension/duckdb_scanner/src/include/duckdb_scanner_extension.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include "extension/extension.h"
-
-namespace kuzu {
-namespace duckdb_scanner {
-
-class DuckDBScannerExtension final : public extension::Extension {
-public:
-    static void load(main::ClientContext* context);
-};
-
-} // namespace duckdb_scanner
-} // namespace kuzu
