--- extension/duckdb_scanner/src/include/duckdb_storage.h
@@ -1,19 +1,0 @@
-#pragma once
-
-#include "main/database.h"
-#include "storage/storage_extension.h"
-
-namespace kuzu {
-namespace duckdb_scanner {
-
-class DuckDBStorageExtension final : public storage::StorageExtension {
-public:
-    static constexpr const char* dbType = "DUCKDB";
-
-    explicit DuckDBStorageExtension(main::Database* database);
-
-    bool canHandleDB(std::string dbType) const override;
-};
-
-} // namespace duckdb_scanner
-} // namespace kuzu
