--- extension/duckdb_scanner/test/test_files/duckdb_scanner.test
@@ -1,140 +1,0 @@
--GROUP DuckDBScanner
--DATASET CSV empty
-
---
-
--CASE ScanDuckDBTable
--STATEMENT load extension "${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/build/libduckdb_scanner.kuzu_extension"
----- ok
--STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/test/duckdb_database/tinysnb.db' as tinysnb (dbtype 'duckdb');
----- 1
-Attached database successfully.
--STATEMENT LOAD FROM tinysnb.person RETURN *;
----- 8
-0|Alice|1|True|False|35|5.000000|1900-01-01|2011-08-20 11:25:30|3 years 2 days 13:02:00|[10,5]|[Aida]|[[10,8],[6,7,8]]|1.731000|a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11|[96,54,86,92]
-10|Hubert Blaine Wolfeschlegelsteinhausenbergerdorff|2|False|True|83|4.900000|1990-11-27|2023-02-21 13:25:30|3 years 2 days 13:02:00|[10,11,12,3,4,5,6,7]|[Ad,De,Hi,Kye,Orlan]|[[7],[10],[6,7]]|1.323000|a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a18|[77,64,100,54]
-2|Bob|2|True|False|30|5.100000|1900-01-01|2008-11-03 15:25:30.000526|10 years 5 months 13:00:00.000024|[12,8]|[Bobby]|[[8,9],[9,10]]|0.990000|a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a12|[98,42,93,88]
-3|Carol|1|False|True|45|5.000000|1940-06-22|1911-08-20 02:32:21|48:24:11|[4,5]|[Carmen,Fred]|[[8,10]]|1.000000|a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a13|[91,75,21,95]
-5|Dan|2|False|True|20|4.800000|1950-07-23|2031-11-30 12:25:30|10 years 5 months 13:00:00.000024|[1,9]|[Wolfeschlegelstein,Daniel]|[[7,4],[8,8],[9]]|1.300000|a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a14|[76,88,99,89]
-7|Elizabeth|1|False|True|20|4.700000|1980-10-26|1976-12-23 11:21:42|48:24:11|[2]|[Ein]|[[6],[7],[8]]|1.463000|a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a15|[96,59,65,88]
-8|Farooq|2|True|False|25|4.500000|1980-10-26|1972-07-31 13:22:30.678559|00:18:00.024|[3,4,5,6,7]|[Fesdwe]|[[8]]|1.510000|a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a16|[80,78,34,83]
-9|Greg|2|False|False|40|4.900000|1980-10-26|1976-12-23 11:21:42|10 years 5 months 13:00:00.000024|[1]|[Grad]|[[10]]|1.600000|a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a17|[43,83,67,43]
--STATEMENT LOAD FROM tinysnb.organisation RETURN *;
----- 3
-1|ABFsUni|325|3.700000|-2|10 years 5 months 13 hours 24 us|3 years 5 days|1.000000|{revenue: 138, "location": ['toronto','montr,eal'], stock: {price: [96,56], volume: 1000}}|3.12
-4|CsWork|934|4.100000|-100|2 years 4 days 10 hours|26 years 52 days 48:00:00|0.780000|{revenue: 152, "location": ["vanco,uver north area"], stock: {price: [15,78,671], volume: 432}}|abcd
-6|DEsWork|824|4.100000|7|2 years 4 hours 22 us 34 minutes|82:00:00.1|0.520000|{revenue: 558, "location": ['very long city name','new york'], stock: {price: [22], volume: 99}}|2023-12-15
--STATEMENT LOAD FROM tinysnb.movies RETURN *;
----- 3
-Roma|298|the movie is very interesting and funny|{rating: 1223.000000, stars: 100, "views": 10003, "release": 2011-02-11 16:44:22, release_ns: 2011-02-11 16:44:22.123456, release_ms: 2011-02-11 16:44:22.123, release_sec: 2011-02-11 16:44:22, release_tz: 2011-02-11 16:44:22.123456+00, film: 2013-02-22, u8: 1, u16: 15, u32: 200, u64: 4, hugedata: -15}|pure ascii characters|{}
-Sóló cón tu párejâ|126| this is a very very good movie|{rating: 5.300000, stars: 2, "views": 152, "release": 2011-08-20 11:25:30, release_ns: 2011-08-20 11:25:30.123456, release_ms: 2011-08-20 11:25:30.123, release_sec: 2011-08-20 11:25:30, release_tz: 2011-08-20 11:25:30.123456+00, film: 2012-05-11, u8: 220, u16: 20, u32: 1, u64: 180, hugedata: 1844674407370955161811111111}|\xAA\xABinteresting\x0B|{audience1=52, audience53=42}
-The 😂😃🧘🏻‍♂️🌍🌦️🍞🚗 movie|2544| the movie is very very good|{rating: 7.000000, stars: 10, "views": 982, "release": 2018-11-13 13:33:11, release_ns: 2018-11-13 13:33:11.123456, release_ms: 2018-11-13 13:33:11.123, release_sec: 2018-11-13 13:33:11, release_tz: 2018-11-13 13:33:11.123456+00, film: 2014-09-12, u8: 12, u16: 120, u32: 55, u64: 1, hugedata: -1844674407370955161511}|\xAB\xCD|{audience1=33}
--STATEMENT LOAD FROM tinysnb.tableOfTypes RETURN count(*);
----- 1
-49999
--STATEMENT LOAD FROM tinysnb.tableOfTypes WHERE id = 49992 or id = 28532 RETURN *;
----- 2
-28532|74|72.472423|True|1977-08-16|TKn|[94,92]|[AUSrJTUWVOESDor,ODOS6RfqMhsFO9aFUa,ziauQj]|[[123,55,181],[32]]|{ID: 666, "name": DiqSQ5u5UhS8aZi}
-49992|50|31.582059|False|2056-05-02||[62,24,94]|[LpQO8OT3x45a]|[[268,281,166],[144,16,126,208,298],[22,287]]|{ID: 936, "name": sGPSafxMAhKiP}
--STATEMENT LOAD FROM tinysnb.person1 RETURN *;
----- error
-Catalog exception: Table: person1 does not exist.
--STATEMENT DETACH tinysnb;
----- 1
-Detached database successfully.
--STATEMENT LOAD FROM tinysnb.person RETURN *;
----- error
-Binder exception: No database named tinysnb has been attached.
--LOG AttachMultipleDuckDB
--STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/test/duckdb_database/tinysnb.db' (dbtype 'duckdb');
----- 1
-Attached database successfully.
--STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/test/duckdb_database/other.db' as Other1 (dbtype 'duckdb');
----- 1
-Attached database successfully.
--STATEMENT LOAD FROM other1.person RETURN *;
----- 4
-1
-2
-3
-5
--STATEMENT LOAD FROM tinysnb.person RETURN count(*);
----- 1
-8
--LOG UseDatabaseStatement
--STATEMENT USE other1;
----- 1
-Used database successfully.
--STATEMENT LOAD FROM person RETURN *;
----- 4
-1
-2
-3
-5
--STATEMENT USE tinysnb;
----- 1
-Used database successfully.
--STATEMENT LOAD FROM person RETURN count(*);
----- 1
-8
--STATEMENT LOAD FROM other1.person RETURN count(*);
----- 1
-4
--STATEMENT CALL DUCKDB_CLEAR_CACHE() RETURN *;
----- ok
--STATEMENT LOAD FROM other1.person RETURN count(*);
----- 1
-4
-
--CASE InvalidDuckDBDatabase
--STATEMENT LOAD FROM tinysnb1.person RETURN *;
----- error
-Binder exception: No database named tinysnb1 has been attached.
--STATEMENT LOAD FROM tinysnb1_person RETURN *;
----- error
-Binder exception: Variable tinysnb1_person is not in scope.
-
--CASE InvalidUseDatabaseName
--STATEMENT USE NONEXIST;
----- error
-Runtime exception: No database named NONEXIST.
-
--CASE ScanDuckdbDatabaseWithoutExt
--STATEMENT load extension "${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/build/libduckdb_scanner.kuzu_extension"
----- ok
--STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/test/duckdb_database/dbfilewithoutext' (dbtype 'duckdb');
----- ok
--STATEMENT LOAD FROM dbfilewithoutext.persons RETURN *
----- 2
-Jeff|32
-Wendy|28
--LOG AttachSameDBError
--STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/test/duckdb_database/dbfilewithoutext' (dbtype 'duckdb');
----- error
-Runtime exception: Database with name: dbfilewithoutext has already been attached.
--LOG DetachExistDB
--STATEMENT DETACH dbfilewithoutext;
----- ok
--LOG DetachNotExistDBError
--STATEMENT DETACH dbfilewithoutext;
----- error
-Runtime exception: Database: dbfilewithoutext doesn't exist.
--LOG AttachNotExistDBError
--STATEMENT ATTACH 'notexist.db' (dbtype 'duckdb');
----- error
-Runtime exception: 'notexist.db' is not a valid duckdb database path.
-
--CASE AttachDBWithoutLoadingExtension
--STATEMENT ATTACH 'notexist.db' (dbtype 'duckdb');
----- error
-Runtime exception: No loaded extension can handle database type: duckdb.
-Did you forget to load duckdb extension?
-You can load it by: load extension duckdb_scanner;
--STATEMENT ATTACH 'notexist.db' (dbtype 'postgres');
----- error
-Runtime exception: No loaded extension can handle database type: postgres.
-Did you forget to load postgres extension?
-You can load it by: load extension postgres_scanner;
--STATEMENT ATTACH 'notexist.db' (dbtype 'sqlite');
----- error
-Runtime exception: No loaded extension can handle database type: sqlite.
