+++ extension/postgres/src/include/postgres_extension.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "extension/extension.h"
+
+namespace kuzu {
+namespace postgres_extension {
+
+class PostgresExtension final : public extension::Extension {
+public:
+    static void load(main::ClientContext* context);
+};
+
+} // namespace postgres_extension
+} // namespace kuzu
