+++ extension/postgres/src/include/postgres_functions.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "function/table_functions.h"
+
+namespace kuzu {
+namespace postgres_extension {
+
+struct PostgresClearCacheFunction final : public function::TableFunction {
+    static constexpr const char* name = "postgres_clear_cache";
+
+    PostgresClearCacheFunction();
+};
+
+} // namespace postgres_extension
+} // namespace kuzu
