+++ extension/postgres/src/include/postgres_storage.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "storage/storage_extension.h"
+
+namespace kuzu {
+namespace postgres_extension {
+
+class PostgresStorageExtension final : public storage::StorageExtension {
+public:
+    static constexpr const char* dbType = "POSTGRES";
+
+    explicit PostgresStorageExtension(main::Database* database);
+
+    bool canHandleDB(std::string dbType) const override;
+};
+
+} // namespace postgres_extension
+} // namespace kuzu
