--- extension/postgres_scanner/src/include/postgres_scanner_extension.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include "extension/extension.h"
-
-namespace kuzu {
-namespace postgres_scanner {
-
-class PostgresScannerExtension final : public extension::Extension {
-public:
-    static void load(main::ClientContext* context);
-};
-
-} // namespace postgres_scanner
-} // namespace kuzu
