+++ test/test_files/extension/show_attached_databases.test
@@ -6,13 +6,13 @@
 -CASE ShowDatabasesTest
 -SKIP_MUSL
 -LOG InstallExtension
--STATEMENT INSTALL duckdb_scanner;
+-STATEMENT INSTALL duckdb;
 ---- ok
--STATEMENT LOAD EXTENSION duckdb_scanner;
+-STATEMENT LOAD EXTENSION duckdb;
 ---- ok
--STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/test/duckdb_database/tinysnb.db' (dbtype 'duckdb');
+-STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb/test/duckdb_database/tinysnb.db' (dbtype 'duckdb');
 ---- ok
--STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb_scanner/test/duckdb_database/dbfilewithoutext' (dbtype 'duckdb');
+-STATEMENT ATTACH '${KUZU_ROOT_DIRECTORY}/extension/duckdb/test/duckdb_database/dbfilewithoutext' (dbtype 'duckdb');
 ---- ok
 -STATEMENT CALL show_attached_databases() RETURN *;
 ---- 2
